+++ src/component/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from './Header';
-import Footer from './Footer';
-import Main from './Main';
+import Header from '.component/Header';
+import Footer from '.component/Footer';
+import Main from '.component/Main';
 
 class App extends React.Component {
   render() {
@@ -14,6 +14,7 @@ class App extends React.Component {
     );
   }
 }
+export default App;
